backend/requirements.txt
# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ML and AI libraries
transformers>=4.35.0
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Audio processing
pydub>=0.25.1
# Note: Whisper is installed separately in setup script for Python 3.14 compatibility
# If setup script fails, try: pip install git+https://github.com/openai/whisper.git

# OpenAI and TTS
openai>=1.3.0
gtts>=2.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0


